REF: Add run command to CLI
