+++ octue/__init__.py
@@ -1,2 +1,2 @@
-from .cli import octue_cli, octue_run  # noqa: F401
+from .cli import octue_cli  # noqa: F401
 from .runner import LOG_FORMAT, Runner  # noqa: F401
